Add flag to ignore private indices

+++ src/planner.js
@@ -41,11 +41,13 @@ const deepEquals = (a: any, b: any): boolean => {
 const plan = (
   newSchema: Schema,
   oldSchema: Schema,
-  hookUrl: ?string
+  hookUrl: ?string,
+  ignorePrivateIndexes: boolean = false,
 ): Array<Command> => {
   return planCollections(
     newSchema.collections,
-    oldSchema.collections
+    oldSchema.collections,
+    ignorePrivateIndexes
   ).concat(
     planFunctions(newSchema.functions, oldSchema.functions, hookUrl),
     planTriggers(newSchema.triggers, oldSchema.triggers, hookUrl)
@@ -54,7 +56,8 @@ const plan = (
 
 const planCollections = (
   newSchema: Array<CollectionDefinition>,
-  oldSchema: Array<CollectionDefinition>
+  oldSchema: Array<CollectionDefinition>,
+  ignorePrivateIndexes: boolean,
 ): Array<Command> => {
   const oldColMap = new Map(oldSchema.map(c => [c.className, c]));
   const newColMap = new Map(newSchema.map(c => [c.className, c]));
@@ -167,6 +170,9 @@ const planCollections = (
       }
       const indexes: { [string]: IndexDefinition } = collection.indexes || {};
       Object.keys(indexes).forEach((name) => {
+        if (ignorePrivateIndexes && name.startsWith('_')) {
+          return;
+        }
         const newIndex = (newC.indexes || {})[name];
         if (newIndex === undefined) {
           di.push(DeleteIndex(collection.className, name));
